// src/model/api/response/styling/karte/CasualPartSizeResponse.ts
export default interface CasualPartSizeResponse {
  readonly shoulder: number | null;
  readonly bust: number | null;
  readonly lengthTop: number | null;
  readonly lengthArm: number | null;
  readonly roundNeck: number | null;
  readonly waist: number | null;
  readonly hip: number | null;
  readonly roundLeg: number | null;
  readonly lengthWaist: number | null;
  readonly lengthLeg: number | null;
  readonly roundCalf: number | null;
  readonly foot: number | null;
  readonly referenceShoulder: number | null;
  readonly referenceBust: number | null;
  readonly referenceLengthTop: number | null;
  readonly referenceLengthArm: number | null;
  readonly referenceWaist: number | null;
  readonly referenceHip: number | null;
  readonly referenceRoundLeg: number | null;
  readonly referenceLengthLeg: number | null;
  readonly referenceRoundCalf: number | null;
}